[Core] Moving OBB
Fixes #9445
