+++ kratos/geometries/oriented_bounding_box.h
