+++ kratos/processes/find_intersected_geometrical_objects_with_obb_process.h
@@ -18,7 +18,7 @@
 // External includes
 
 // Project includes
-#include "includes/oriented_bounding_box.h"
+#include "geometries/oriented_bounding_box.h"
 #include "processes/find_intersected_geometrical_objects_process.h"
 
 namespace Kratos
